planning-agent/requirements.txt
mcp>=1.0.0
httpx>=0.25.0
tenacity>=8.0.0
structlog>=24.0.0
python-dotenv>=1.0.0